Implementation of the optimal givens rotation network.
