+++ openfermioncirq/primitives/__init__.py
@@ -14,6 +14,8 @@
 
 from openfermioncirq.primitives.bogoliubov_transform import bogoliubov_transform
 
+from openfermioncirq.primitives.optimal_givens_decomposition import front_and_back_givens_decomposition
+
 from openfermioncirq.primitives.state_preparation import (
     prepare_gaussian_state,
     prepare_slater_determinant)
